Fixed the trait
Having the getters and setters for secret here will hide them from access by base class' uri builder. Result is a barcode without secret which is invalid.

Please be aware that 'sha1' is hardcoded here - a bad idea, but needs a bit more work. Do not change the algorithm in the config file without adjusting it here too.

Please be also aware that Google Authenticator cannot handle e.g. sha256 - you'll be safe for the moment if you use the default sha1